pyproject.toml
[project]
name = "snapfiction-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.94.0",
    "google-genai>=1.16.1",
    "loguru>=0.7.3",
    "moviepy==1.0.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
]
